requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
jinja2==3.1.4
httpx==0.27.2
docker==7.1.0
paramiko==3.5.0
apscheduler==3.10.4
pyyaml==6.0.2
python-multipart==0.0.12
aiofiles==24.1.0
python-dotenv==1.0.1
